requirements.txt
# Core LangChain & LangGraph
langchain>=0.1.20
langgraph>=0.0.40
langchain-core>=0.1.40
langchain-community>=0.0.38
langchain-huggingface>=0.0.1

# LLM Providers
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
langchain-groq>=0.1.3
langgraph-checkpoint-postgres>=2.0.0
psycopg[binary,pool]>=3.2.0

# Vector Store & Embeddings
pinecone-client>=3.2.2
sentence-transformers>=2.2.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.9.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Backend API
fastapi
uvicorn
python-multipart

# Optional: Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
